Refactor code generator to use templates
With the exception of switch/case ordering in GetDocForID and
AddDataElement; this change doesn't make changes to the generated code.

This refactor does a number of things:
- moves code generation from string manipulation to Go's text/templates.
- consolidates each generated function into a distinct template
- separates the main function into 3 distinct steps gathering data,
assembling data to use for code generation, and then code generation.
- To support the separation of main, it splits existing fill functions
into distinct steps, focused on the template that it is providing data
for.
- sorts the case ordering in the switch statements in GetDocForID and
AddDataElement.

This change improves maintainability of the generator function, mainly
by separating concerns & centralizing code generation.