Update generator & linetypes to preserve type info
Now that we have structs to represent a full document, we can switch
addDataElement to be a method on each struct. This change:
- Updates code generation ordering to reflect that addDataElement is a
method of the parent document struct
- Adds a METdocument interface to util so we can easily refer to any
document struct
- Preserves type information in our map of ID keys & document structs

This introduces a change to the signature of ParseLine() - we now
require our map of docs to be of type map[string]METdocument.