Add a GetID method to our document structs
We need to be able to easily extract the ID from a METdocument
interface. We want to keep the list of methods in METdocument minimal.
However, we may also want to allow LINE_TYPE extraction so we can more
easily determine which type we have, especially if we wanted to avoid
reflection.