Refactor util.GetID() to BuildID() and return a string
util.GetID() was previously setting the ID field on the passed-in
VxMetadata struct and returning a pointer to the updated struct. This
made usage unclear.

Returning the ID string makes the purpose of BuildID clear.

This could have been a method on the VxMetadata struct. However, it
requires that other fields in VxMetadata be correctly set so requiring
them as part of the function signature here seemed clearer.